sibilant importer will be enabled immediately from now on. this means the import order needs adjusting
