--- sibilant/importer.py
@@ -1,37 +1,0 @@
-# This library is free software; you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as
-# published by the Free Software Foundation; either version 3 of the
-# License, or (at your option) any later version.
-#
-# This library is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public
-# License along with this library; if not, see
-# <http://www.gnu.org/licenses/>.
-
-
-"""
-sibilant.importer
-
-Importing this module will augment the python importing system to
-be able to load sibilant. Calls sibilant.importlib.install()
-
-author: Christopher O'Brien  <obriencj@gmail.com>
-license: LGPL v.3
-"""
-
-
-from .importlib import install
-
-
-__all__ = ()
-
-
-install()
-
-
-#
-# The end.
